Saturday update
implementation of a query builder. The query builder can create joins, update, and selects. With a custom query params. Now it's easy to consume sql server. But we need some tests to improve that this is really working. Now we need to refactor all sql queries to use the Query Builder function. We need to implmenet the manager function to get clients-vip relation. I made a mistake on the interpretation of this requeriment.