All controllers are tested properly and return the result expected.
Today we do some bug fix on CepValidator and Export the postman
collection. Changed the targetapi.sql script to create the database and
insert the vips fields in VipModel table. We fix all broken fields. But
we still needs fix Swagger config to solve the opcional params problem.
This is a small issue. But, at this moment all queries and all routes
are true working and tested. But.. about the database, i still think
we need a table with all relational fields independent of all others
table. This will improve performace the queries in sql server. But.. now
is too late to do this critical change on Database.