Pre finalization of API Target Desafio v1
This is last commit. This project is my first contact with swagger on
aspcore framework. I read many articles to gets all this api works i
how was expected. But.. i dont implemented all the cool functions of swagger
here. I will try to implement the cool features of swagger on my own tests.
But here we have a native conection with Microsoft SQL Server, a query
builder algorithm that i will try to implement sooner or later in my
nexts personal projects, we have a small MVCS project schema here. I
tried to implement a S.O.L.I.D concept here but i think the concept of Liskov
is not successfull implemented in all classes. But.. this is all i did here.